+++ packages/swarm-components/lib/index.js
@@ -3,12 +3,6 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-Object.defineProperty(exports, "NumericalInput", {
-  enumerable: true,
-  get: function get() {
-    return _NumericalInput.default;
-  }
-});
 Object.defineProperty(exports, "Button", {
   enumerable: true,
   get: function get() {
@@ -27,10 +21,28 @@ Object.defineProperty(exports, "Checkbox", {
     return _Checkbox.default;
   }
 });
+Object.defineProperty(exports, "FieldHelper", {
+  enumerable: true,
+  get: function get() {
+    return _FieldHelper.default;
+  }
+});
+Object.defineProperty(exports, "FieldLabel", {
+  enumerable: true,
+  get: function get() {
+    return _FieldLabel.default;
+  }
+});
 Object.defineProperty(exports, "Icon", {
   enumerable: true,
   get: function get() {
     return _Icon.default;
+  }
+});
+Object.defineProperty(exports, "NumericalInput", {
+  enumerable: true,
+  get: function get() {
+    return _NumericalInput.default;
   }
 });
 Object.defineProperty(exports, "Radio", {
@@ -70,15 +82,19 @@ Object.defineProperty(exports, "Select", {
   }
 });
 
-var _NumericalInput = _interopRequireDefault(require("./NumericalInput"));
-
 var _Button = _interopRequireDefault(require("./Button"));
 
 var _LinkButton = _interopRequireDefault(require("./LinkButton"));
 
 var _Checkbox = _interopRequireDefault(require("./Checkbox"));
 
+var _FieldHelper = _interopRequireDefault(require("./FieldHelper"));
+
+var _FieldLabel = _interopRequireDefault(require("./FieldLabel"));
+
 var _Icon = _interopRequireDefault(require("./Icon"));
+
+var _NumericalInput = _interopRequireDefault(require("./NumericalInput"));
 
 var _Radio = _interopRequireDefault(require("./Radio"));