+++ packages/swarm-components/src/index.js
@@ -1,8 +1,8 @@
-import NumericalInput from './NumericalInput';
-
 export { default as Button } from './Button';
 export { default as LinkButton } from './LinkButton';
 export { default as Checkbox } from './Checkbox';
+export { default as FieldHelper } from './FieldHelper';
+export { default as FieldLabel } from './FieldLabel';
 export { default as Icon } from './Icon';
 export { default as NumericalInput } from './NumericalInput';
 export { default as Radio } from './Radio';