+++ packages/swarm-docs/CHANGELOG.md
@@ -3,6 +3,97 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.16](https://github.com/gatsbyjs/gatsby-starter-default/compare/@meetup/swarm-docs@0.7.15...@meetup/swarm-docs@0.7.16) (2019-04-02)
+
+
+### Bug Fixes
+
+* adding select docs with FieldLabel and FieldHelper ([e467290](https://github.com/gatsbyjs/gatsby-starter-default/commit/e467290))
+
+
+
+
+
+## [0.7.15](https://github.com/gatsbyjs/gatsby-starter-default/compare/@meetup/swarm-docs@0.7.14...@meetup/swarm-docs@0.7.15) (2019-04-02)
+
+**Note:** Version bump only for package @meetup/swarm-docs
+
+
+
+
+
+## [0.7.14](https://github.com/gatsbyjs/gatsby-starter-default/compare/@meetup/swarm-docs@0.7.13...@meetup/swarm-docs@0.7.14) (2019-04-02)
+
+**Note:** Version bump only for package @meetup/swarm-docs
+
+
+
+
+
+## [0.7.13](https://github.com/gatsbyjs/gatsby-starter-default/compare/@meetup/swarm-docs@0.7.12...@meetup/swarm-docs@0.7.13) (2019-04-02)
+
+**Note:** Version bump only for package @meetup/swarm-docs
+
+
+
+
+
+## [0.7.12](https://github.com/gatsbyjs/gatsby-starter-default/compare/@meetup/swarm-docs@0.7.11...@meetup/swarm-docs@0.7.12) (2019-04-01)
+
+**Note:** Version bump only for package @meetup/swarm-docs
+
+
+
+
+
+## [0.7.11](https://github.com/gatsbyjs/gatsby-starter-default/compare/@meetup/swarm-docs@0.7.10...@meetup/swarm-docs@0.7.11) (2019-04-01)
+
+**Note:** Version bump only for package @meetup/swarm-docs
+
+
+
+
+
+## [0.7.10](https://github.com/gatsbyjs/gatsby-starter-default/compare/@meetup/swarm-docs@0.7.9...@meetup/swarm-docs@0.7.10) (2019-04-01)
+
+**Note:** Version bump only for package @meetup/swarm-docs
+
+
+
+
+
+## [0.7.9](https://github.com/gatsbyjs/gatsby-starter-default/compare/@meetup/swarm-docs@0.7.8...@meetup/swarm-docs@0.7.9) (2019-04-01)
+
+**Note:** Version bump only for package @meetup/swarm-docs
+
+
+
+
+
+## [0.7.8](https://github.com/gatsbyjs/gatsby-starter-default/compare/@meetup/swarm-docs@0.7.7...@meetup/swarm-docs@0.7.8) (2019-04-01)
+
+**Note:** Version bump only for package @meetup/swarm-docs
+
+
+
+
+
+## [0.7.7](https://github.com/gatsbyjs/gatsby-starter-default/compare/@meetup/swarm-docs@0.7.6...@meetup/swarm-docs@0.7.7) (2019-04-01)
+
+**Note:** Version bump only for package @meetup/swarm-docs
+
+
+
+
+
+## [0.7.6](https://github.com/gatsbyjs/gatsby-starter-default/compare/@meetup/swarm-docs@0.7.5...@meetup/swarm-docs@0.7.6) (2019-04-01)
+
+**Note:** Version bump only for package @meetup/swarm-docs
+
+
+
+
+
 ## [0.7.5](https://github.com/gatsbyjs/gatsby-starter-default/compare/@meetup/swarm-docs@0.7.4...@meetup/swarm-docs@0.7.5) (2019-03-29)
 
 **Note:** Version bump only for package @meetup/swarm-docs